staterecovery: decouple request retry and logging in blobscan client
